refactor: split code into modules for constants, events, game logic, game over, and UI
- Moved constants, event handling, game logic, game over, and UI rendering into separate modules: constants.py, event_handler.py, game_logic.py, game_over.py, and ui_renderer.py.
- Improved project organization for better maintainability and readability.
